requirements.txt
annotated-types==0.7.0
anyio==4.12.1
aiohttp==3.10.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
playwright==1.57.0
playwright-stealth==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyshorteners==1.0.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
soupsieve==2.8.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
